style: rename dto 'NotificationResponse' -> 'NotificationRegisterResponse'
