+++ src/main/java/com/wanted/preonboarding/ticket/application/notification/PerformanceNotificationRegister.java
@@ -3,7 +3,7 @@ package com.wanted.preonboarding.ticket.application.notification;
 import com.wanted.preonboarding.ticket.domain.notification.Notification;
 import com.wanted.preonboarding.ticket.domain.notification.NotificationRepository;
 import com.wanted.preonboarding.ticket.domain.performance.PerformanceRepository;
-import com.wanted.preonboarding.ticket.dto.response.alarm.NotificationResponse;
+import com.wanted.preonboarding.ticket.dto.response.notification.NotificationRegisterResponse;
 import com.wanted.preonboarding.ticket.exception.notfound.NotFoundException;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
@@ -26,7 +26,7 @@ public class PerformanceNotificationRegister implements NotificationRegister {
 
     @Transactional
     @Override
-    public NotificationResponse register(final String targetId, final String name, final String phone) {
+    public NotificationRegisterResponse register(final String targetId, final String name, final String phone) {
         if (!isTargetExist(targetId)) throw new NotFoundException("알람 대상을 찾을 수 없습니다.");
 
         Notification notification = Notification.builder()
@@ -35,6 +35,6 @@ public class PerformanceNotificationRegister implements NotificationRegister {
             .phoneNumber(phone)
             .build();
 
-        return NotificationResponse.of(alarmRepository.save(notification));
+        return NotificationRegisterResponse.of(alarmRepository.save(notification));
     }
 }
